Update product.rb
Group-by update is required for tech maturity to work with a PostgreSQL data store (otherwise the result is a syntax error: `ERROR:  column "tags.id" must appear in the GROUP BY clause or be used in an aggregate function`). This change does not appear to affect functionality or correctness for SQLite-backed instances of the app.